+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.107.0 (Unreleased)
+FEATURES:
+* mongodb: added `enable_flow_control`, `compressors`, `prefix_compression`, `slow_op_sample_rate` parameters to mongodb config
+
 
 ## 0.106.0 (January 23, 2024)
 FEATURES: